Refactor HunyuanImage3 image processing helpers
