+++ vllm_omni/diffusion/models/hunyuan_image3/pipeline_hunyuan_image3.py
@@ -49,6 +49,7 @@ from .hunyuan_image3_transformer import (
     real_batched_index_select,
     retrieve_timesteps,
 )
+from .image_processing import resize_and_crop
 from .system_prompt import get_system_prompt
 
 if TYPE_CHECKING:
@@ -127,20 +128,7 @@ def _resize_and_crop_center(image: PILImage.Image, target_width: int, target_hei
     # Mirrors HunyuanImage3Processor._resize_and_crop in
     # vllm_omni.model_executor.models.hunyuan_image3.hunyuan_image3 so the AR
     # and DiT stages preprocess condition images identically.
-    tw, th = target_width, target_height
-    w, h = image.size
-    tr = th / tw
-    r = h / w
-    if r < tr:
-        resize_height = th
-        resize_width = int(round(th / h * w))
-    else:
-        resize_width = tw
-        resize_height = int(round(tw / w * h))
-    resized = image.resize((resize_width, resize_height), PILImage.Resampling.LANCZOS)
-    crop_top = int(round((resize_height - th) / 2.0))
-    crop_left = int(round((resize_width - tw) / 2.0))
-    return resized.crop((crop_left, crop_top, crop_left + tw, crop_top + th))
+    return resize_and_crop(image, (target_width, target_height), crop_type="center")
 
 
 def _to_python_scalar(value: Any) -> Any:
